requirements.txt
unittest-xml-reporting
xcodestream==0.0.8
nltk==3.4.5
pandas==0.25.1
azure-storage-blob
apache-airflow
requests
pyshorteners==0.6.1
Munch==2.3.2
pyyaml==5.1.1
memoization